Correction cloud 3

+++ main.py
@@ -20,7 +20,7 @@ def main(input, res, octave):
         os.mkdir('output/')
 
     tifffile.imwrite('output/background.png', u.astype(np.uint16))
-    tifffile.imwrite('output/cloud.png', (cloud * 255).astype(np.uint16))
+    tifffile.imwrite('output/cloud.png', (cloud * 255).astype(np.uint8))
     tifffile.imwrite('output/cloudy.png', cloudy.astype(np.uint16))
